31 integrate task tracker api with master api

+++ GemConnectAPI/Controllers/TaskTracker/TaskTrackerController.cs
@@ -1,18 +1,23 @@
 ﻿using Task = TaskTrackerConsole.model.Task;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Authorization;
 using TaskTrackerConsole.data;
 using TaskTrackerConsole.model;
-using TaskTrackerConsole.dto;
+using System.Security.Claims;
+using GemConnectAPI.Mappers.TaskTracker;
+using GemConnectAPI.DTO.TaskTracker;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
-namespace TaskTrackerAPI.Controllers
+namespace GemConnectAPI.Controllers.TaskTracker
 {
+    [Authorize(Policy = "USER")]
     [Route("api/[controller]")]
     [ApiController]
     public class TaskTrackerController : ControllerBase
     {
         private readonly ITaskManager _taskManager;
+        private readonly ITaskMapper _taskMapper;
         private readonly ILogger<TaskTrackerController> _logger;
 
         /// <summary>
@@ -21,9 +26,11 @@ namespace TaskTrackerAPI.Controllers
         /// <param name="taskManager">The task manager service.</param>
         /// <param name="logger">The logger service.</param>
 
-        public TaskTrackerController(ITaskManager taskManager, ILogger<TaskTrackerController> logger)
+        public TaskTrackerController(ITaskManager taskManager, ITaskMapper taskMapper,
+            ILogger<TaskTrackerController> logger)
         {
             _taskManager = taskManager;
+            _taskMapper = taskMapper;
             _logger = logger;
         }
         /// <summary>
@@ -37,11 +44,16 @@ namespace TaskTrackerAPI.Controllers
         ///
         /// </remarks>
         [HttpGet("tasks")]
-        public ActionResult<IEnumerable<TaskDto>> Get()
+        public ActionResult<IEnumerable<TaskDto>> GetTasks()
         {
-            var tasks = _taskManager.GetTasks();
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;  // Get the User Id from the 
+
+            // Fetch tasks that belong to the authenticated user
+            var tasks = _taskManager.GetTasks().Where(t => t.CreatedBy == int.Parse(userId));
+
+            var username = User.FindFirst(ClaimTypes.Name)?.Value;  // Fetch the username
             //Return the tasks as TaskDto objects instead of Task models
-            IEnumerable<TaskDto> enumerable = tasks.Select(MapTaskToDto);  //pass the method MapTaskToDto to the Select method
+            IEnumerable<TaskDto> enumerable = tasks.Select(task => _taskMapper.MapTaskToDto(task, username));  //pass the method MapTaskToDto to the Select method
             var taskDtos = enumerable;
 
             return Ok(taskDtos);
@@ -61,7 +73,7 @@ namespace TaskTrackerAPI.Controllers
         ///
         /// </remarks>
         [HttpGet("task/{id}")]
-        public IActionResult Get(int id)
+        public IActionResult GetTaskById(int id)
         {
             var task = _taskManager.GetTask(id);
 
@@ -69,8 +81,15 @@ namespace TaskTrackerAPI.Controllers
             {
                 return HandleTaskNotFound(id);
             }
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
 
-            var taskDto = MapTaskToDto(task);
+            // Ensure the user is the creator of the task or is an Admin
+            if (task.CreatedBy != int.Parse(userId) && !User.IsInRole("ADMIN"))
+            {
+                return Forbid("You are not allowed to view this task.");
+            }
+            var username = User.FindFirst(ClaimTypes.Name)?.Value;
+            TaskDto taskDto = _taskMapper.MapTaskToDto(task, username);
 
             return Ok(taskDto);
         }
@@ -92,37 +111,33 @@ namespace TaskTrackerAPI.Controllers
         ///     }
         ///
         /// </remarks>
+        /// // <summary>
+        /// Creates a new task (User-level access).
+        /// </summary>
+        [Authorize(Policy = "USER")]  // Allow users to create tasks
         [HttpPost]
-        public IActionResult Post([FromBody] TaskDto taskDto)
+        public IActionResult AddTask([FromBody] CreateTaskDto taskDto)
         {
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
             }
+            // Parse and validate the status here
 
-            // Try to parse the string status to the Status enum
-            if (!TryParseStatus(taskDto.Status, out Status statusEnum))  // Corrected to use "!"
-            {
-                _logger.LogError($"ERROR: Invalid status value '{taskDto.Status}' provided.");
-                return BadRequest("Invalid status value. Please provide a valid status like 'TODO', 'PENDING', or 'COMPLETE'.");
-            }
 
-            Task newTask = new()
-            {
-                Description = taskDto.Description,
-                Status = statusEnum,
-                CreatedAt = DateTime.UtcNow,
-                UpdatedAt = DateTime.UtcNow
-            };
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;  // Get the User Id from the token
+            //Validate and parse the status
+            var (statusEnum, errorResponse) = ValidateAndParseStatus(taskDto.Status);
+            if (errorResponse != null) { return errorResponse; }
+
+            // Map to Task using the mapper, now passing the already validated statusEnum
+            Task newTask = _taskMapper.MaptoTask(taskDto, int.Parse(userId), statusEnum.Value);
 
             // Add the new task
             _taskManager.AddTask(newTask);
-
-            // Use the helper method to map to TaskDto
-            var newTaskDto = MapTaskToDto(newTask);
-
             // Return 201 Created response with the new task's URI
-            return CreatedAtAction(nameof(Get), new { id = newTask.Id }, newTaskDto);
+            // return CreatedAtAction(nameof(GetTaskById), new { id = newTask.Id });
+            return Ok("New task created.");
         }
 
         /// <summary>
@@ -145,8 +160,9 @@ namespace TaskTrackerAPI.Controllers
         ///
         /// </remarks>
         // PUT api/<TaskTrackerController>/5
+        [Authorize(Policy = "USER")]
         [HttpPut("{id}")]
-        public IActionResult Put(int id, [FromBody] TaskDto taskDto)
+        public IActionResult UpdateTask(int id, [FromBody] CreateTaskDto taskDto)
         {
             var task = _taskManager.GetTask(id);
 
@@ -154,20 +170,27 @@ namespace TaskTrackerAPI.Controllers
             {
                 return HandleTaskNotFound(id);
             }
-            // Try to parse the string status to the Status 
-            if (!TryParseStatus(taskDto.Status, out Status statusEnum))
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+
+            // Ensure the user is the creator of the task or is an Admin
+            if (task.CreatedBy != int.Parse(userId) && !User.IsInRole("ADMIN"))
             {
-                _logger.LogError(message: $"ERROR: Invalid status value '{taskDto.Status}' provided.");
-                return BadRequest("Invalid status value. Please provide a valid status like 'TODO', 'PENDING', or 'COMPLETE'.");
+                return Forbid("You are not allowed to view this task.");
             }
+            // Try to parse the string status to the Status 
+            var (statusEnum, errorResponse) = ValidateAndParseStatus(taskDto.Status);
+            if (errorResponse != null) { return errorResponse; }
+
             task.Description = taskDto.Description;
-            task.Status = statusEnum;
+            task.Status = (Status)statusEnum;
             task.UpdatedAt = DateTime.UtcNow; //Update the timestamp
 
             //Update task
             _taskManager.UpdateTask(id, task);
+            var username = User.FindFirst(ClaimTypes.Name)?.Value;
+
             // Use the helper method to map to TaskDto
-            var updatedTaskDto = MapTaskToDto(task);
+            TaskDto updatedTaskDto = _taskMapper.MapTaskToDto(task, username: username);
             return Ok(updatedTaskDto);
         }
 
@@ -185,6 +208,7 @@ namespace TaskTrackerAPI.Controllers
         ///
         /// </remarks>
         // DELETE api/<TaskTrackerController>/5
+        [Authorize(Policy = "USER")]
         [HttpDelete("{id}")]
         public IActionResult Delete(int id)
         {
@@ -193,33 +217,44 @@ namespace TaskTrackerAPI.Controllers
             {
                 return HandleTaskNotFound(id);
             }
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
 
-            //Delete task
+            // Ensure the user is the creator of the task or is an Admin
+            if (task.CreatedBy != int.Parse(userId) && !User.IsInRole("ADMIN"))
+            {
+                return Forbid("You are not allowed to view this task.");
+            }
+            // Ensure users can only delete their own tasks or if they are Admins
+            if (task.CreatedBy != int.Parse(userId) && !User.IsInRole("ADMIN"))
+            {
+                return Forbid("You are not allowed to delete this task.");
+            }
+
             _taskManager.DeleteTask(id);
             return NoContent();
-
         }
 
         #region Helpers
-        private IActionResult HandleTaskNotFound(int id)
+        private (Status?, IActionResult) ValidateAndParseStatus(string statusString)
+        {
+            // Attempt to parse the status string to the enum
+            if (Enum.TryParse(statusString, true, out Status statusEnum))
+            {
+                // Parsing successful, return the statusEnum and no error response
+                return (statusEnum, null);
+            }
+
+            // Log the error and return a BadRequest response with the error message
+            _logger.LogError($"Invalid status value '{statusString}' provided.");
+            return (null, BadRequest("Invalid status value. Please provide a valid status like 'TODO', 'PENDING', or 'COMPLETE'."));
+        }
+        private NotFoundObjectResult HandleTaskNotFound(int id)
         {
             // Use a logger for better logging management (assumes ILogger<TaskTrackerController> is injected)
             _logger.LogWarning(message: $"Task with ID {id} could not be found.");
 
             // Return a 404 Not Found response
             return NotFound($"Task with ID {id} could not be found.");
-        }
-        private bool TryParseStatus(string statusString, out Status status) => Enum.TryParse(statusString, true, out status);
-        private TaskDto MapTaskToDto(Task task)
-        {
-            return new TaskDto
-            {
-                Id = task.Id,
-                Description = task.Description,
-                Status = task.Status.ToString(),
-                CreatedAt = task.CreatedAt.ToString("yyyy-MM-dd HH:mm:ss"),
-                UpdatedAt = task.UpdatedAt.ToString("yyyy-MM-dd HH:mm:ss")
-            };
         }
         #endregion
     }
